codes/015-blog-site/requirements.txt
Django==2.2.3
django-autoslug==1.9.4
django-jet==1.0.8
pytz==2019.1
sqlparse==0.3.0
